+++ fields/components/DateInput.js
@@ -1,5 +1,5 @@
 import moment from 'moment';
-import DayPicker, { DateUtils } from 'react-day-picker';
+import DayPicker from 'react-day-picker';
 import React from 'react';
 import { findDOMNode } from 'react-dom';
 import Popout from '../../admin/client/App/shared/Popout';
@@ -14,11 +14,7 @@ module.exports = React.createClass({
 		name: React.PropTypes.string,
 		onChange: React.PropTypes.func.isRequired,
 		path: React.PropTypes.string,
-		range: React.PropTypes.bool,
-		value: React.PropTypes.oneOfType([
-		    React.PropTypes.string,
-		    React.PropTypes.object,
-		]),
+		value: React.PropTypes.string,
 	},
 	getDefaultProps () {
 		return {
@@ -28,24 +24,13 @@ module.exports = React.createClass({
 	getInitialState () {
 		const id = ++lastId;
 		let month = new Date();
-		var { value } = this.props;
-		const { format, range } = this.props;
-		if (range) {
-			if (value.from && moment(value.from, format, true).isValid()) {
-				month = moment(value.from, format).toDate();
-			}
-			value = value || {
-				from: undefined,
-				to: undefined,
-			};
-		} else {
-			if (moment(value, format, true).isValid()) {
-				month = moment(value, format).toDate();
-			}
+		const { format, value } = this.props;
+		if (moment(value, format, true).isValid()) {
+			month = moment(value, format).toDate();
 		}
 		return {
 			id: `_DateInput_${id}`,
-			month,
+			month: month,
 			pickerIsOpen: false,
 			inputValue: value,
 		};
@@ -54,18 +39,9 @@ module.exports = React.createClass({
 		this.showCurrentMonth();
 	},
 	componentWillReceiveProps: function (newProps) {
-		let month = new Date();
-		if (this.props.range) {
-			if (JSON.stringify(newProps.value) === JSON.stringify(this.props.value)) return;
-			if (newProps.value.from) {
-				month = moment(newProps.value.from, this.props.format).toDate();
-			}
-		} else {
-			if (newProps.value === this.props.value) return;
-			month = moment(newProps.value, this.props.format).toDate();
-		}
+		if (newProps.value === this.props.value) return;
 		this.setState({
-			month,
+			month: moment(newProps.value, this.props.format).toDate(),
 			inputValue: newProps.value,
 		}, this.showCurrentMonth);
 	},
@@ -80,43 +56,25 @@ module.exports = React.createClass({
 	handleKeyPress (e) {
 		if (e.key === 'Enter') {
 			e.preventDefault();
-			if (this.props.range) {
-				const { inputValue: { from: fromDate, to: toDate } } = this.state;
-				if (fromDate && moment(fromDate, this.props.format, true).isValid() 
-					&& toDate && moment(toDate, this.props.format, true).isValid()) {
-					this.props.onChange({ value: this.state.inputValue });
-				// If the date is not strictly equal, only change the tab that is displayed
-				} else if (fromDate && moment(fromDate, this.props.format).isValid() 
-					&& toDate && moment(toDate, this.props.format).isValid()) {
-					this.setState({
-						month: moment(fromDate, this.props.format).toDate(),
-					}, this.showCurrentMonth);
-				}
-			} else {
-				// If the date is strictly equal to the format string, dispatch onChange
-				if (moment(this.state.inputValue, this.props.format, true).isValid()) {
-					this.props.onChange({ value: this.state.inputValue });
-				// If the date is not strictly equal, only change the tab that is displayed
-				} else if (moment(this.state.inputValue, this.props.format).isValid()) {
-					this.setState({
-						month: moment(this.state.inputValue, this.props.format).toDate(),
-					}, this.showCurrentMonth);
-				}
+			// If the date is strictly equal to the format string, dispatch onChange
+			if (moment(this.state.inputValue, this.props.format, true).isValid()) {
+				this.props.onChange({ value: this.state.inputValue });
+			// If the date is not strictly equal, only change the tab that is displayed
+			} else if (moment(this.state.inputValue, this.props.format).isValid()) {
+				this.setState({
+					month: moment(this.state.inputValue, this.props.format).toDate(),
+				}, this.showCurrentMonth);
 			}
 		}
 	},
-	handleDaySelect (date) {
-		// if (modifiers && modifiers.disabled) return;
-		// console.log(date);
-		let value = null;
-		if (this.props.range) {
-			value = DateUtils.addDayToRange(date, this.state.inputValue);
-		} else {
-			value = moment(date).format(this.props.format)
-		}
+	handleDaySelect (e, date, modifiers) {
+		if (modifiers && modifiers.disabled) return;
+
+		var value = moment(date).format(this.props.format);
+
 		this.props.onChange({ value });
 		this.setState({
-			pickerIsOpen: !!this.props.range,
+			pickerIsOpen: false,
 			month: date,
 			inputValue: value,
 		});
@@ -126,7 +84,7 @@ module.exports = React.createClass({
 	},
 	showCurrentMonth () {
 		if (!this.refs.picker) return;
-		// this.refs.picker.showMonth(this.state.month);
+		this.refs.picker.showMonth(this.state.month);
 	},
 	handleFocus (e) {
 		if (this.state.pickerIsOpen) return;
@@ -138,47 +96,20 @@ module.exports = React.createClass({
 	handleBlur (e) {
 		let rt = e.relatedTarget || e.nativeEvent.explicitOriginalTarget;
 		const popout = this.refs.popout.getPortalDOMNode();
-		console.log(popout);
 		while (rt) {
 			if (rt === popout) return;
 			rt = rt.parentNode;
 		}
-		console.log(rt);
 		this.setState({
 			pickerIsOpen: false,
 		});
 	},
 	render () {
-		const { range, format } = this.props;
-		var { value } = this.props;
-		var { inputValue } = this.state;
-		var modifiers;
-		var selectedDay;
-		// if (range) {
-		// 	// selectedDay = { from: value.from, to: value.to };
-		if (range) {
-			selectedDay = { from: value.from, to: value.to };
-			modifiers = { 
-				selectedRange: {
-					from: value.from,
-					to: value.to,
-				},
-			};
-			// if (range) {
-			if (inputValue.from && inputValue.to) {
-				inputValue = `${inputValue.from.toLocaleDateString()} - ${inputValue.to.toLocaleDateString()}`;
-			} else if (inputValue.from) {
-				inputValue = `${inputValue.from.toLocaleDateString()}`;
-			} else if (inputValue.to) {
-				inputValue = `${inputValue.to.toLocaleDateString()}`;
-			}
-		} else {
-			selectedDay = this.props.value;
-			// react-day-picker adds a class to the selected day based on this
-			modifiers = {
-				selected: (day) => moment(day).format(this.props.format) === selectedDay,
-			};
-		}
+		const selectedDay = this.props.value;
+		// react-day-picker adds a class to the selected day based on this
+		const modifiers = {
+			selected: (day) => moment(day).format(this.props.format) === selectedDay,
+		};
 		var { maxDate, minDate } = this.props;
 		var optional = {};
 		if (maxDate) {
@@ -202,17 +133,9 @@ module.exports = React.createClass({
 				},
 			};
 		}
-		console.log('range: ', this.props);
+
 		return (
 			<div>
-				{
-					range && <FormInput
-						autoComplete="off"
-						name={this.props.name}
-						style={{ display: 'none' }}
-						value={inputValue}
-					/>
-				}
 				<FormInput
 					autoComplete="off"
 					id={this.state.id}
@@ -222,28 +145,22 @@ module.exports = React.createClass({
 					onFocus={this.handleFocus}
 					onKeyPress={this.handleKeyPress}
 					placeholder={this.props.format}
-					placeholder={
-						range ? `${format} - ${format}` : format}
-					style={
-						range ? { width: '230px' } : null
-					}
 					ref="input"
-					value={inputValue}
+					value={this.state.inputValue}
 				/>
 				<Popout
 					isOpen={this.state.pickerIsOpen}
 					onCancel={this.handleCancel}
 					ref="popout"
 					relativeToID={this.state.id}
-					width={range ? 530 : 260}
-				>
+					width={260}
+					>
 					<DayPicker
-						className="Selectable"
 						modifiers={modifiers}
-						numberOfMonths={range ? 2 : 1}
 						onDayClick={this.handleDaySelect}
 						ref="picker"
 						tabIndex={-1}
+						
 					/>
 				</Popout>
 			</div>