+++ fields/types/daterange/DateRangeColumn.js
@@ -8,14 +8,19 @@ var DateRangeColumn = React.createClass({
 	propTypes: {
 		col: React.PropTypes.object,
 		data: React.PropTypes.object,
+		format: React.PropTypes.string,
 		linkTo: React.PropTypes.string,
 	},
 	getValue () {
 		const value = this.props.data.fields[this.props.col.path];
 		if (!value) return null;
-
-		const format = (this.props.col.type === 'datetime') ? 'HH:mm DD/MM/YYYY' : 'MMMM Do YYYY';
-		return moment(value).format(format);
+		const { format } = this.props;
+		const dateFormat = (this.props.col.type === 'datetime') ? 
+			(format ? format : 'DD/MM/YYYY HH:mm') : (format ? format : 'DD/MM/YYYY');
+		if (value.startDate && value.endDate) {
+			return `${moment(value.startDate).format(dateFormat)} - ${moment(value.endDate).format(dateFormat)}`;
+		}
+		return '';
 	},
 	render () {
 		const value = this.getValue();